+++ Engine/Code/CoreTools/Base/SpanIteratorDetail.h
@@ -383,7 +383,7 @@ void CoreTools::SpanIterator<Iter>::SetValue(int step, T value)
 }
 
 template <typename Iter>
-template <typename T, int Size>
+template <typename T, size_t Size>
 requires(sizeof(typename Iter::value_type) <= sizeof(T) && sizeof(T) % sizeof(typename Iter::value_type) == 0)
 void CoreTools::SpanIterator<Iter>::SetValue(int step, const std::array<T, Size>& value)
 {
